+++ web/app/react/components/application.component.jsx
@@ -1,4 +1,4 @@
-import * as React from 'react';
+import React from 'react';
 import Header from './header.component.jsx';
 
 class Application extends React.Component {